Release version 0.8.0 (#433)
* add last modified by and last modified at to script information (#337)

* add action type to action result table (#332)

* add created by & created at in script information (#335)

* store all design actions in the database (#336)

* fix spelling mistake for http.executeRequest (#346)

* Connection Creation - HTTP Connection Type - Missing Mandatory Field (#347)

* bump version

* Allow proxy parameter of http.executeRequest to be Null (#351)

* update circleci configuration

* update circleci configuration

* handle semantic versioning when loading OpenAPI Documentation (#355)

* allow execution requests to be filtered on run id (#358)

* add UUID instruction (#360)

* add instruction for number formatting (#359)

* add pagination to environment endpoints (#352)

* complete database creation creation for IESI connections (#306)

* store raw response in output dataset of http.execute response (#370)

* bugfix: update script with same version (#376)

* allow filtering on script execution status for execution requests (#362)

* make unit tests client unspecific (#381)

* start scripts as separate processes (#367)

* Allow users to use specific version of a component for http execute request (#380)

* Refactor access token logic to contain only roles (#282)




* wip

* Revert "wip"

This reverts commit a7f647cf5b90827bc6e94c828ae55acaac66deaf.

* update sql statement when fetching latest component version

* Allows user to connect to a SQL database from any provider by creating a generic connection type (#382)

* Allow users to request multiple script executions (#384)

* Add encryption api endpoint to encrypt values (#389)

Added encryption API endpoint to encrypt values

* Allows user to assign components to a security group and protect sensitive data from other groups

* Allows user to assign components to a security group and protect sensitive data from other groups

* Allows user to assign datasets to a security group and protect sensitive data from other groups

* removed local repository.db3 file

* removed local yml changes

* Math instructions fixed and working

* Enhancement of API response (#397)

Enhancement of API response

* Make requestVersion optional from action http.executeRequest (#404)

* Fix the log4j2 vulnerability

* Allow users to manage datasets through the UI

* Allow users to apply preferences when setting headers while executing a http request

* add missing files that should come from the branch headers definition preferences

* Make action parameter appendOutput optional

* add additional conditition to convertAppendOutput and convertDatasetReference name methods from action sql.executeQuery

* Remove exit property from script execution request

* Revert "Allow users to apply preferences when setting headers while executing a http request" (#413)

This reverts commit 852cf5f9b9c14ad51409be8b7a37594da6a52f64.

* Revert "add missing files that should come from the branch headers definition preferences"

This reverts commit f3800bb2700fac0154e7f3d0fa9290bb5f3cc344.

* Add xml/json path traversal as instruction

* Define clob length of action parameter value column to 4000

* Fix bug on instructions that are not properly handled when multiple closing braces appear (#420)

* Define clob length of action parameter value column to 4000

* Fix issue when the script status is getting blocked on "running" during json path traversal

* Allow users to store action outputs to in memory datasets in order to use temporary the results within a script execution

* Implement new syntax for the http headers

* removes spaces before checking the regex pattern in the instruction text.xmlpath

* fix multiple instruction and headers issues

* add InMemoryDatasetService and DatabaseDatasetService in DataTypeHandler

* Bump xercesImpl from 2.12.0 to 2.12.2 in /core/java/iesi-core (#426)

* update 'configure your script using subroutines' section (#373)

* Bump xercesImpl from 2.12.0 to 2.12.2 in /core/java/iesi-core

Bumps xercesImpl from 2.12.0 to 2.12.2.

---
updated-dependencies:
- dependency-name: xerces:xercesImpl
  dependency-type: direct:production
...



* Bump jackson-databind from 2.9.6 to 2.9.10.8 in /core/java/iesi-cockpit/iesi-cockpit-ui (#406)

* update 'configure your script using subroutines' section (#373)

* Bump jackson-databind in /core/java/iesi-cockpit/iesi-cockpit-ui

Bumps [jackson-databind](https://github.com/FasterXML/jackson) from 2.9.6 to 2.9.10.8.
- [Release notes](https://github.com/FasterXML/jackson/releases)
- [Commits](https://github.com/FasterXML/jackson/commits)

---
updated-dependencies:
- dependency-name: com.fasterxml.jackson.core:jackson-databind
  dependency-type: direct:production
...



* Adding bugfix for the xml path instructions and spaces preservations. Adding instruction escapements + small adjustments

* bugfix xml path instruction, spaces preservation, escaping instruction lookup concept

* add instruction escapments

* small ajdustments

* handling of runtimes variables which are not properly closed

* assign one repository coordinator for each metadata when type general is provided on the conf file

* if input is only null as text, return new Null()

+++ core/java/iesi-rest-without-microservices/src/main/java/io/metadew/iesi/server/rest/component/ComponentsController.java
@@ -3,21 +3,26 @@ package io.metadew.iesi.server.rest.component;
 import io.metadew.iesi.metadata.configuration.exception.MetadataAlreadyExistsException;
 import io.metadew.iesi.metadata.configuration.exception.MetadataDoesNotExistException;
 import io.metadew.iesi.metadata.definition.component.key.ComponentKey;
+import io.metadew.iesi.metadata.service.user.IESIPrivilege;
 import io.metadew.iesi.metadata.tools.IdentifierTools;
 import io.metadew.iesi.server.rest.component.dto.ComponentDto;
 import io.metadew.iesi.server.rest.component.dto.ComponentDtoResourceAssembler;
 import io.metadew.iesi.server.rest.component.dto.IComponentDtoService;
+import io.metadew.iesi.server.rest.configuration.security.IesiSecurityChecker;
 import io.metadew.iesi.server.rest.error.DataBadRequestException;
 import io.metadew.iesi.server.rest.resource.HalMultipleEmbeddedResource;
-import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PagedResourcesAssembler;
 import org.springframework.hateoas.PagedModel;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PostAuthorize;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -29,35 +34,39 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 
 @RestController
-@Tag(name = "components", description = "Everything about components")
 @RequestMapping("/components")
+@ConditionalOnWebApplication
 public class ComponentsController {
 
     private final IComponentService componentService;
     private final IComponentDtoService componentDtoService;
     private final ComponentDtoResourceAssembler componentDtoResourceAssembler;
     private final PagedResourcesAssembler<ComponentDto> componentDtoPagedResourcesAssembler;
+    private final IesiSecurityChecker iesiSecurityChecker;
 
     @Autowired
     ComponentsController(ComponentDtoResourceAssembler componentDtoResourceAssembler,
                          IComponentService componentService,
                          IComponentDtoService componentDtoService,
-                         PagedResourcesAssembler<ComponentDto> componentDtoPagedResourcesAssembler) {
+                         PagedResourcesAssembler<ComponentDto> componentDtoPagedResourcesAssembler,
+                         IesiSecurityChecker iesiSecurityChecker
+    ) {
         this.componentDtoResourceAssembler = componentDtoResourceAssembler;
         this.componentService = componentService;
         this.componentDtoService = componentDtoService;
         this.componentDtoPagedResourcesAssembler = componentDtoPagedResourcesAssembler;
+        this.iesiSecurityChecker = iesiSecurityChecker;
     }
 
     @GetMapping("")
     @PreAuthorize("hasPrivilege('COMPONENTS_READ')")
     public PagedModel<ComponentDto> getAll(Pageable pageable, @RequestParam(required = false, name = "name") String name) {
         List<ComponentFilter> componentFilters = extractComponentFilterOptions(name);
-        Page<ComponentDto> componentDtoPage = componentDtoService.getAll(pageable, componentFilters);
+        Page<ComponentDto> componentDtoPage = componentDtoService.getAll(SecurityContextHolder.getContext().getAuthentication(), pageable, componentFilters);
 
         if (componentDtoPage.hasContent())
             return componentDtoPagedResourcesAssembler.toModel(componentDtoPage, componentDtoResourceAssembler::toModel);
-        //noinspection unchecked
+
         return (PagedModel<ComponentDto>) componentDtoPagedResourcesAssembler.toEmptyModel(componentDtoPage, ComponentDto.class);
     }
 
@@ -72,35 +81,35 @@ public class ComponentsController {
     @GetMapping("/{name}")
     @PreAuthorize("hasPrivilege('COMPONENTS_READ')")
     public PagedModel<ComponentDto> getByName(Pageable pageable, @PathVariable String name) {
-        Page<ComponentDto> componentDtoPage = componentDtoService.getByName(pageable, name);
+        Page<ComponentDto> componentDtoPage = componentDtoService.getByName(SecurityContextHolder.getContext().getAuthentication(), pageable, name);
         if (componentDtoPage.hasContent())
             return componentDtoPagedResourcesAssembler.toModel(componentDtoPage, componentDtoResourceAssembler::toModel);
-        //noinspection unchecked
+
         return (PagedModel<ComponentDto>) componentDtoPagedResourcesAssembler.toEmptyModel(componentDtoPage, ComponentDto.class);
     }
 
     @GetMapping("/{name}/{version}")
     @PreAuthorize("hasPrivilege('COMPONENTS_READ')")
+    @PostAuthorize("hasPrivilege('COMPONENTS_READ', returnObject.securityGroupName)")
     public ComponentDto get(@PathVariable String name, @PathVariable Long version) throws MetadataDoesNotExistException {
-        ComponentDto component = componentDtoService.getByNameAndVersion(name, version)
+        return componentDtoService.getByNameAndVersion(null, name, version)
                 .orElseThrow(() -> new MetadataDoesNotExistException(new ComponentKey(IdentifierTools.getComponentIdentifier(name), version)));
-        return componentDtoResourceAssembler.toModel(component);
     }
 
     @PostMapping("")
-    @PreAuthorize("hasPrivilege('COMPONENTS_WRITE')")
-    public ComponentDto post(@Valid @RequestBody ComponentDto component) {
+    @PreAuthorize("hasPrivilege('COMPONENTS_WRITE', #componentDto.securityGroupName)")
+    public ComponentDto post(@Valid @RequestBody ComponentDto componentDto) {
         try {
-            componentService.createComponent(component);
-            return componentDtoResourceAssembler.toModel(component.convertToEntity());
+            componentService.createComponent(componentDto);
+            return componentDtoResourceAssembler.toModel(componentDtoService.convertToEntity(componentDto));
         } catch (MetadataAlreadyExistsException e) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND,
-                    "Component " + component.getName() + " already exists");
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
+                    "Component " + componentDto.getName() + " already exists");
         }
     }
 
     @PutMapping("")
-    @PreAuthorize("hasPrivilege('COMPONENTS_WRITE')")
+    @PreAuthorize("hasPrivilege('COMPONENTS_WRITE', #componentDtos.![securityGroupName])")
     public HalMultipleEmbeddedResource<ComponentDto> putAll(@Valid @RequestBody List<ComponentDto> componentDtos) throws MetadataDoesNotExistException {
         componentService.updateComponents(componentDtos);
         HalMultipleEmbeddedResource<ComponentDto> halMultipleEmbeddedResource = new HalMultipleEmbeddedResource<>();
@@ -115,36 +124,31 @@ public class ComponentsController {
     }
 
     @PutMapping("/{name}/{version}")
-    @PreAuthorize("hasPrivilege('COMPONENTS_WRITE')")
-    public ComponentDto put(@PathVariable String name, @PathVariable Long version, @RequestBody ComponentDto component) throws MetadataDoesNotExistException {
-        if (!component.getName().equals(name)) {
+    @PreAuthorize("hasPrivilege('COMPONENTS_WRITE', #componentDto.securityGroupName)")
+    public ComponentDto put(@PathVariable String name, @PathVariable Long version, @RequestBody ComponentDto componentDto) throws MetadataDoesNotExistException {
+        if (!componentDto.getName().equals(name)) {
             throw new DataBadRequestException(name);
-        } else if (component.getVersion().getNumber() != version) {
+        } else if (componentDto.getVersion().getNumber() != version) {
             throw new DataBadRequestException(version);
         }
-        componentService.updateComponent(component);
-        return componentDtoResourceAssembler.toModel(component.convertToEntity());
+        componentService.updateComponent(componentDto);
+        return componentDtoResourceAssembler.toModel(componentDtoService.convertToEntity(componentDto));
 
-    }
-
-    @DeleteMapping("")
-    @PreAuthorize("hasPrivilege('COMPONENTS_WRITE')")
-    public ResponseEntity<?> deleteAll() {
-        componentService.deleteAll();
-        return ResponseEntity.status(HttpStatus.OK).build();
-    }
-
-    @DeleteMapping("/{name}")
-    @PreAuthorize("hasPrivilege('COMPONENTS_WRITE')")
-    public ResponseEntity<?> deleteByName(@PathVariable String name) {
-        componentService.deleteByName(name);
-        return ResponseEntity.status(HttpStatus.OK).build();
     }
 
     @DeleteMapping("/{name}/{version}")
     @PreAuthorize("hasPrivilege('COMPONENTS_WRITE')")
     public ResponseEntity<?> delete(@PathVariable String name, @PathVariable Long version) throws MetadataDoesNotExistException {
+        ComponentDto componentDto = componentDtoService.getByNameAndVersion(null, name, version)
+                .orElseThrow(() -> new MetadataDoesNotExistException(new ComponentKey(IdentifierTools.getComponentIdentifier(name), version)));
+        if (!iesiSecurityChecker.hasPrivilege(SecurityContextHolder.getContext().getAuthentication(),
+                IESIPrivilege.COMPONENTS_MODIFY.getPrivilege(),
+                componentDto.getSecurityGroupName())
+        ) {
+            throw new AccessDeniedException("User is not allowed to delete this component");
+        }
         componentService.deleteByNameAndVersion(name, version);
+
         return ResponseEntity.status(HttpStatus.OK).build();
     }
 }
